+++ src/dataset/__init__.py
@@ -1,0 +1,3 @@
+
+from .chessboard_ext import chessboard_to_compact_2d_feature_maps as conv_board
+from .gm_games_dataset import ChessGmGamesDataset
