+++ src/legacy/README.md
@@ -1,0 +1,4 @@
+
+## Legacy Content
+Those were the first naive training approaches.
+This might be deleted on future releases.
