+++ src/pre-train/__init__.py
@@ -1,0 +1,5 @@
+
+from .pretrain_models import ChessFeatureExtractionModel, ChessDrawGenerator, ChessRatingModel
+from .pretrain_feature_ext import DrawGenTrainingSession
+from .pretrain_rating import RatingTrainingSession
+from .pretrain_model_learing import DeepQTrainingSession
